requirements.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=3.15
whitenoise>=6.5
gunicorn>=21.2
python-dotenv>=1.0
mysqlclient>=2.2
django-filter>=23.5
djangorestframework-simplejwt>=5.3
django-crispy-forms>=2.0
drf-yasg>=1.21
dj-database-url>=2.1
django-import-export>=3.3